flask_backtest_app/requirements.txt
Flask==2.3.2
PyYAML==6.0
openbb
pandas
cachetools
python-dotenv